+++ src/chat/frontend/chat-root/components/chat-input/components/control-panel/control-panel.ts
@@ -20,7 +20,6 @@ export class ControlPanel extends LitElement {
     @property({ type: Boolean }) isSummarizing = false;
     @property({ type: Number }) agentToken = 0;
     @property({ type: Boolean }) hasTokenData = false;
-    @property({ type: Boolean }) isGhost = false;
     @property({ type: Array }) contextSearchResults: ContextSearchResult[] = [];
     @property({ type: Boolean }) autoAccept = false;
 
@@ -93,7 +92,6 @@ export class ControlPanel extends LitElement {
                     .isSummarizing=${this.isSummarizing}
                     .agentToken=${this.agentToken}
                     .hasTokenData=${this.hasTokenData}
-                    .isGhost=${this.isGhost}
                     .contextSearchResults=${this.contextSearchResults}
                     .autoAccept=${this.autoAccept}
                     @submit-click=${this._handleSubmit}